pyproject.toml
[tool.poetry]
name = "omtool"
version = "0.2.0"
description = "Package and program that models N-Body problem in galactic evolution application."
authors = ["Artyom Zaporozhets <kraysent@gmail.com>"]
license = "Apache-2.0"
readme = "README.rst"
repository = "https://github.com/Kraysent/OMTool"
documentation = "https://omtool.readthedocs.io/en/latest/"

[tool.poetry.dependencies]
python = "^3.10"
amuse-framework = "^2022.6.0"
astropy = "^5.1"
numpy = "^1.23.1"
matplotlib = "^3.5.2"
marshmallow = "^3.17.0"
marshmallow-jsonschema = "^0.13.0"
PyYAML = "^6.0"
pandas = "^1.4.3"
pyzerolog = "^0.3.0"
# pyfalcon = { git = "https://github.com/GalacticDynamics-Oxford/pyfalcon.git" }
py-expression-eval = { version = "^0.3.14", optional = true }

[tool.poetry.dev-dependencies]
black = "^22.6.0"
flake8 = "^4.0.1"
isort = "^5.10.1"
mypy = "^0.961"
types-PyYAML = "^6.0.10"

[tool.poetry.extras]
tasks = ["py-expression-eval"]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
